+++ src/packages/user/user/invite/entity-action/resend-invite/resend-invite.action.condition.ts
@@ -1,4 +1,4 @@
-import { UmbUserActionConditionBase } from './user-allow-action-base.condition.js';
+import { UmbUserActionConditionBase } from '../../../conditions/user-allow-action-base.condition.js';
 import { UserStateModel } from '@umbraco-cms/backoffice/backend-api';
 import {
 	ManifestCondition,
