+++ src/packages/user/user/invite/manifests.ts
@@ -1,5 +1,11 @@
 import { manifests as collectionActionManifests } from './collection-action/manifests.js';
 import { manifests as modalManifests } from './modal/manifests.js';
 import { manifests as repositoryManifests } from './repository/manifests.js';
+import { manifests as entityActionManifests } from './entity-action/manifests.js';
 
-export const manifests = [...collectionActionManifests, ...modalManifests, ...repositoryManifests];
+export const manifests = [
+	...collectionActionManifests,
+	...modalManifests,
+	...repositoryManifests,
+	...entityActionManifests,
+];
